pyproject.toml
[project]
name = "sync-image-gen"
version = "0.1.1"
description = "Sync directory monitor with Nano Banana Pro (Gemini) image processing and full-screen display"
readme = "README.md"
authors = [
    { name = "Derek Hsu" }
]
license = { text = "MIT" }
keywords = ["gemini", "nano-banana", "image-processing", "directory-monitor", "watchdog"]
requires-python = ">=3.10"
dependencies = [
    "watchdog>=4.0.0",
    "google-genai>=0.1.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/derekhsu/sync-image-gen"

[project.scripts]
sync-image-gen = "sync_image_gen:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
